+++ arklet-tunnel-mqtt/src/main/java/com/alipay/sofa/koupleless/arklet/tunnel/mqtt/paho/PahoMqttClient.java
@@ -189,7 +189,7 @@ public class PahoMqttClient {
         public void messageArrived(String topic, MqttMessage mqttMessage) throws Exception {
             // message callback, don't pub sync message here, may cause deadlock
             String[] topicSplits = topic.split("/");
-            messageHandler.handle(topicSplits[topicSplits.length - 1], mqttMessage);
+            messageHandler.handleCommand(topicSplits[topicSplits.length - 1], mqttMessage);
         }
 
         @Override
@@ -197,199 +197,5 @@ public class PahoMqttClient {
             // send message successfully, no former process here
         }
 
-    }
-
-    public static class MqttMessageHandler {
-
-        private final CommandService   commandService;
-        private final BaseMetadataHook baseMetadataHook;
-        private final MqttClient       mqttClient;
-        private final UUID             deviceID;
-        private String                 baseEnv;
-        private final AtomicBoolean    run = new AtomicBoolean(false);
-
-        public MqttMessageHandler(CommandService commandService, BaseMetadataHook baseMetadataHook,
-                                  MqttClient mqttClient, UUID deviceID, String baseEnv) {
-            this.commandService = commandService;
-            this.baseMetadataHook = baseMetadataHook;
-            this.mqttClient = mqttClient;
-            this.deviceID = deviceID;
-            this.baseEnv = baseEnv;
-            if (this.baseEnv == null || this.baseEnv.isEmpty()) {
-                this.baseEnv = Constants.DEFAULT_BASE_ENV;
-            } else {
-                this.baseEnv = this.baseEnv.toLowerCase();
-            }
-        }
-
-        /**
-         * <p>getHealthTopic.</p>
-         *
-         * @return String
-         */
-        private String getHealthTopic() {
-            return String.format("koupleless_%s/%s/base/health", baseEnv, deviceID);
-        }
-
-        /**
-         * <p>getHeartBeatTopic.</p>
-         *
-         * @return String
-         */
-        private String getHeartBeatTopic() {
-            return String.format("koupleless_%s/%s/base/heart", baseEnv, deviceID);
-        }
-
-        /**
-         * <p>getBizTopic.</p>
-         *
-         * @return String
-         */
-        private String getBizTopic() {
-            return String.format("koupleless_%s/%s/base/biz", baseEnv, deviceID);
-        }
-
-        /**
-         * <p>getBizTopic.</p>
-         *
-         * @return String
-         */
-        private String getBizOperationResponseTopic() {
-            return String.format("koupleless_%s/%s/base/bizOperation", baseEnv, deviceID);
-        }
-
-        /**
-         * <p>getCommandTopic.</p>
-         *
-         * @return String
-         */
-        private String getCommandTopic() {
-            return String.format("koupleless_%s/%s/+", baseEnv, deviceID);
-        }
-
-        static class HeartBeatScheduledMission implements Runnable {
-
-            private final String           topic;
-            private final MqttClient       mqttClient;
-            private final CommandService   commandService;
-            private final BaseMetadataHook baseMetadataHook;
-
-            public HeartBeatScheduledMission(String topic, MqttClient mqttClient,
-                                             CommandService commandService,
-                                             BaseMetadataHook baseMetadataHook) {
-                this.topic = topic;
-                this.mqttClient = mqttClient;
-                this.commandService = commandService;
-                this.baseMetadataHook = baseMetadataHook;
-            }
-
-            @Override
-            public void run() {
-                // send heart beat message
-                Map<String, Object> heartBeatData = new HashMap<>();
-
-                BaseMetadata metadata = baseMetadataHook.getBaseMetadata();
-                heartBeatData.put(
-                    com.alipay.sofa.koupleless.arklet.core.health.model.Constants.MASTER_BIZ_INFO,
-                    metadata);
-
-                Map<String, String> networkInfo = new HashMap<>();
-
-                try {
-                    InetAddress localHost = InetAddress.getLocalHost();
-                    networkInfo.put(Constants.LOCAL_IP, localHost.getHostAddress());
-                    networkInfo.put(Constants.LOCAL_HOST_NAME, localHost.getHostName());
-                } catch (UnknownHostException e) {
-                    throw new ArkletRuntimeException("get local host failed", e);
-                }
-
-                heartBeatData.put(Constants.NETWORK_INFO, networkInfo);
-                heartBeatData.put(Constants.State, BizState.ACTIVATED);
-
-                try {
-                    mqttClient.publish(topic,
-                        JSONObject.toJSONString(MqttResponse.withData(heartBeatData)).getBytes(), 1,
-                        false);
-                } catch (MqttException e) {
-                    LOGGER.info("mqtt client publish health status failed");
-                    throw new ArkletRuntimeException("mqtt client publish health status failed", e);
-                }
-            }
-        }
-
-        public void run() {
-            try {
-                mqttClient.subscribe(getCommandTopic(), 1);
-            } catch (MqttException e) {
-                LOGGER.info("mqtt client subscribe command topic failed");
-                throw new ArkletInitException("mqtt client subscribe command topic failed", e);
-            }
-            if (run.compareAndSet(false, true)) {
-                ScheduledExecutorService executor = Executors.newScheduledThreadPool(1);
-
-                executor.scheduleAtFixedRate(new HeartBeatScheduledMission(getHeartBeatTopic(),
-                    mqttClient, commandService, baseMetadataHook), 0, 120000L,
-                    TimeUnit.MILLISECONDS);
-            }
-        }
-
-        public void handle(String cmd, MqttMessage msg) throws Exception {
-            // process mqtt message, use thread pool to handle command
-            ThreadPoolExecutor executor = ExecutorServiceManager.getArkTunnelMqttExecutor();
-            executor.submit(() -> {
-                Map<String, Object> cmdContent = JSONObject.parseObject(msg.toString(),
-                    HashMap.class);
-                Output<?> output;
-                try {
-                    // process the command
-                    output = commandService.process(cmd, cmdContent);
-                } catch (InterruptedException e) {
-                    LOGGER.error("process command failed", e);
-                    throw new ArkletRuntimeException(e);
-                }
-
-                try {
-                    if (cmd.equals(BuiltinCommand.HEALTH.getId())) {
-                        // health command, send result to health topic
-                        mqttClient.publish(getHealthTopic(),
-                            JSONObject.toJSONString(MqttResponse.withData(output)).getBytes(), 1,
-                            false);
-                    } else if (cmd.equals(BuiltinCommand.QUERY_ALL_BIZ.getId())) {
-                        // queryAllBiz command, send result to biz topic
-                        mqttClient.publish(getBizTopic(),
-                            JSONObject.toJSONString(MqttResponse.withData(output)).getBytes(), 1,
-                            false);
-                    } else {
-                        // install or uninstall command, send result to biz operation response topic when failed
-                        if (output.failed()) {
-                            Map<String, Object> bizOperationResponse = new HashMap<>();
-                            bizOperationResponse.put(Constants.COMMAND, cmd);
-                            bizOperationResponse.put(Constants.COMMAND_RESPONSE, output);
-                            mqttClient.publish(getBizOperationResponseTopic(),
-                                JSONObject.toJSONString(MqttResponse.withData(bizOperationResponse))
-                                    .getBytes(),
-                                1, false);
-                        } else {
-                            // biz operation, need to sync biz status, queryAllBiz and send result to biz topic
-                            Output<?> allBizOutput;
-                            try {
-                                allBizOutput = commandService
-                                    .process(BuiltinCommand.QUERY_ALL_BIZ.getId(), null);
-                            } catch (InterruptedException e) {
-                                LOGGER.error("queryAllBiz command process failed", e);
-                                throw new ArkletRuntimeException(e);
-                            }
-                            mqttClient.publish(
-                                getBizTopic(), JSONObject
-                                    .toJSONString(MqttResponse.withData(allBizOutput)).getBytes(),
-                                1, false);
-                        }
-                    }
-                } catch (MqttException e) {
-                    LOGGER.error("mqtt publish failed", e);
-                    throw new ArkletRuntimeException(e);
-                }
-            });
-        }
     }
 }
